refactor(core): partial of [Refactoring] Landmark version `1.0.0-alpha.5` #62
