+++ avilla/core/protocol.py
@@ -9,7 +9,7 @@ from avilla.core.account import AbstractAccount
 
 # from avilla.core.action.extension import ActionExtension
 # from avilla.core.action.middleware import ActionMiddleware
-from avilla.core.context import ctx_avilla, ctx_protocol
+from avilla.core._runtime import ctx_avilla, ctx_protocol
 from avilla.core.event import AvillaEvent
 from avilla.core.platform import Abstract, Land, Platform
 from avilla.core.querier import ProtocolAbstractQueryHandler
