+++ avilla/core/querier.py
@@ -17,7 +17,7 @@ from avilla.core.utilles.selector import Selector
 
 if TYPE_CHECKING:
     from avilla.core.protocol import BaseProtocol
-    from avilla.core.relationship import Context
+    from avilla.core.context import Context
 
 
 def query(target: str):
