+++ avilla/core/tools/filter.py
@@ -10,7 +10,7 @@ from graia.broadcast import BaseDispatcher, DispatcherInterface, ExecutionStop
 from typing_extensions import ParamSpec, Self
 
 from avilla.core.account import AbstractAccount
-from avilla.core.relationship import Context
+from avilla.core.context import Context
 from avilla.core.utilles import classproperty
 from avilla.core.utilles.selector import Selectable, Selector
