+++ avilla/core/trait/recorder.py
@@ -28,7 +28,7 @@ from .context import ctx_prefix, eval_dotpath, get_current_namespace
 
 if TYPE_CHECKING:
     from avilla.core.metadata import Metadata, MetadataRoute
-    from avilla.core.relationship import Context
+    from avilla.core.context import Context
     from avilla.core.trait.signature import ArtifactSignature
 
     from . import Fn
