+++ avilla/elizabeth/account.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from avilla.core.account import AbstractAccount
-from avilla.core.relationship import Context
+from avilla.core.context import Context
 from avilla.core.utilles.selector import Selector
 from avilla.elizabeth.connection import ElizabethConnection
 from avilla.elizabeth.connection.util import CallMethod
